sparate Todo
