+++ component/TodoList.js
@@ -2,46 +2,18 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import {
   FlatList,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  View,
 } from 'react-native';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    flexDirection: 'row',
-    alignContent: 'center',
-  },
-  item: {
-    padding: 10,
-    fontSize: 18,
-  },
-  delete: {
-    margin: 4,
-    padding: 10,
-    marginLeft: 'auto',
-    backgroundColor: '#f00',
-  },
-  deleteText: {
-    color: '#fff',
-  },
-});
+import Todo from './Todo';
 
 class TodoList extends Component {
   _renderItem = ({ item }) => {
     const { deleteItem } = this.props;
+
     return (
-      <View style={styles.container}>
-        <Text style={styles.item}>{item.text}</Text>
-        <TouchableOpacity
-          style={styles.delete}
-          onPress={() => deleteItem(item)}
-        >
-          <Text style={styles.deleteText}>DONE</Text>
-        </TouchableOpacity>
-      </View>
+      <Todo
+        item={item}
+        deleteItem={deleteItem}
+      />
     );
   };
